refactor: split src/ into apps/popup, apps/options-ready shared/
- apps/popup/: App.tsx, App.css, index.tsx, components/ (Header, Content, Footer)
- shared/: components/ (AddressList, ClickToCopyText, FeedbackPopover), hooks/, states/, models/, storage.ts, utils.ts
- vite.config.ts: @popup/@options/@shared aliases replace @/*, no generic root alias
- tsconfig.app.json, package.json lint script, index.html script path updated to match
- tests/ reorganized to mirror (tests/apps/popup/, tests/shared/)
- zero behavior change - same 34 tests pass at their new locations
