+++ apps/popup/components/Footer.tsx
