+++ index.html
@@ -10,6 +10,6 @@
 </head>
 <body>
   <div id="root"></div>
-  <script type="module" src="/src/index.tsx"></script>
+  <script type="module" src="/apps/popup/index.tsx"></script>
 </body>
 </html>
