+++ package.json
@@ -13,7 +13,7 @@
     "build": "tsc -b && vite build",
     "dev": "vite",
     "preview": "vite preview",
-    "lint": "eslint src tests",
+    "lint": "eslint apps shared tests",
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage",
