+++ shared/components/AddressList/index.tsx
@@ -9,11 +9,10 @@ import styled from "styled-components";
 /**
  * Internal modules
  */
-// import "./AddressList.css";
-import type { AddressData } from "@/shared/models/address";
+import type { AddressData } from "@shared/models/address";
 import { ClickToCopyText } from "../ClickToCopyText";
 import { FeedbackPopover } from "../FeedbackPopover";
-import type { DisplayOptions } from "@/shared/models/settings";
+import type { DisplayOptions } from "@shared/models/settings";
 
 interface Props extends DisplayOptions {
   data: AddressData[];
