+++ shared/components/ClickToCopyText.tsx
