+++ shared/components/index.ts
@@ -1,0 +1,2 @@
+export { AddressList } from "./AddressList";
+export { ClickToCopyText } from "./ClickToCopyText";
