+++ shared/storage.ts
