--- src/components/index.ts
@@ -1,2 +1,0 @@
-export { AddressList } from "./ui/AddressList";
-export { ClickToCopyText } from "./ui/ClickToCopyText";
