+++ vite-env.d.ts
